fix: auditoria completa pre-producao — 10 correcoes
Frontend:
- Recepcao.jsx: bug data de nascimento — mensagem de erro agora some
  assim que o usuario digita uma data valida (calcularIdade retorna objeto)
- Recepcao.jsx: modo edicao reseta `registrado=false` ao entrar,
  evitando botao desabilitado indevidamente

Backend:
- recepcao_service.py: remover() usava repo.get() (sem carregar relacao);
  corrigido para get_by_id() consistente com o resto do service
- schemas/paciente.py: PacienteUpdate.nacionalidade era Optional=None,
  podendo gerar erro 500 no banco (coluna NOT NULL); agora default=010
- models/recepcao_atendimento.py: BigInteger -> Integer para id e
  paciente_id, consistente com pacientes.id (Integer)

Migracao (migrate_to_postgres.py):
- DDL pacientes: VARCHAR(50) -> VARCHAR(100) em nome/mae/logpcn/numpcn/
  bairro/nome_social/cidade/ocupacao/responsavel; logpcn/numpcn/
  bairro_pcnte agora nullable (modelo SQLAlchemy); adiciona coluna
  naturalidade VARCHAR(100); num_cpf UNIQUE
- DDL recepcao_atendimentos: adiciona colunas faltantes que o backend
  usa — classificacao_risco (enum), observacoes, historia_clinica,
  hipotese_diagnostica, created_at, updated_at; renomeia criado_em
- Cria enum classificacao_risco_enum antes das tabelas (IF NOT EXISTS)
- _COLS_PAC: adiciona naturalidade (inserida como NULL)
- transformar_pac: remove defaults forcados (PRINCIPAL/S-N/CENTRO);
  aumenta limites de truncamento para 100 chars

Scripts/Deploy:
- backup_postgres.bat: BACKUP_DIR -> C:\HMPCF\backups; mkdir antes
  de usar; data via PowerShell Get-Date (locale-safe); limpeza via
  Get-ChildItem ao inves de forfiles (mais portavel)
- DEPLOY_HMPCF_REMOTO.ps1: remove $migResult inutilizado; localIP com
  null-check (SEM-IP-CONFIGURADO como fallback); bat gerado com as
  mesmas correcoes do backup_postgres.bat

+++ backend/app/models/recepcao_atendimento.py
@@ -2,7 +2,7 @@ import enum
 from datetime import datetime
 from typing import TYPE_CHECKING, Optional
 
-from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, Index, SmallInteger, Text, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Index, Integer, SmallInteger, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database.base import Base
@@ -27,10 +27,10 @@ class RecepcaoAtendimento(Base):
 
     __tablename__ = "recepcao_atendimentos"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     paciente_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("pacientes.id", ondelete="RESTRICT"),
         nullable=False,
     )
